+++ packages/shared/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @sugarat/theme-shared
 
+## 0.0.3
+
+### Patch Changes
+
+- feat: add getVitePressPages
+
 ## 0.0.2
 
 ### Patch Changes