+++ packages/theme/CHANGELOG.md
@@ -5,8 +5,13 @@
 ### Patch Changes
 
 - feat: inline group icons plugin
+- feat: support i18n
+- feat: support rewrites and dynamic routes
 - chore: Optimization of theme reloading logic in dev mode
-
+- Updated dependencies
+  - @sugarat/theme-shared@0.0.3
+  - vitepress-plugin-pagefind@0.4.11
+  - vitepress-plugin-rss@0.2.11
 ## 0.4.13
 
 ### Patch Changes